requirements.txt
arrow==1.2.2
certifi==2021.5.30
cycler==0.11.0
Cython==0.29.28
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
Jinja2==3.0.3
jinja2-time==0.2.0
kiwisolver==1.3.1
make==0.1.6.post2
MarkupSafe==2.0.1
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.5
olefile==0.46
Pillow==8.4.0
pycocotools-windows==2.0.0.2
pyparsing==3.0.8
python-dateutil==2.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
wincertstore==0.2
